+++ src/Domain/Model/HoaDonTienDienVN.java
@@ -14,7 +14,7 @@ public class HoaDonTienDienVN extends HoaDonTienDien {
 
     private doiTuongKH doiTuong;
 
-    public HoaDonTienDienVN(int idKh, String hoTen, Date ngayHD, int soLuong, double donGia, double dinhMuc,
+    public HoaDonTienDienVN(int idKh, String hoTen, Date ngayHD, double soLuong, double donGia, double dinhMuc,
             double thanhTien) {
         super(idKh, hoTen, ngayHD, soLuong, donGia, thanhTien);
         this.dinhMuc = dinhMuc;
